fewer spurious errors when creating objects with computed props
Summary:
In a discussion with SamChou19815 he had the idea of moving the code that deals with objects with
computed properties outside of Flow_js. Indeed, `CreateObjWithComputedPropT` no longer needs to
be a constraint, since its inputs should be already resolved under LTI.

The main benefit of this move is that it allows types to be concretized and thus cleaned up, before they are
processed as a computed property. This allows us to not error on code like
```
function test(key: ?string) {
  if (key == null) {
    return;
  }
  const obj = { [key]: "" }; // okay: union boils away with concretization
}
```
Before this code would error because the type of key was a union of `empty | string`, where the `empty`
originated from refining the `?` part.

This diffs does this by moving the respective code to statement.ml. Before we can inspect the types, we
introduce a concretization step. We use a new concretization target `ConcretizeComputedPropsT`,
instead of reusing the existing ones, that is placed around the same location where
`CreateObjWithComputedPropT` was before, since we want the same transformations on the input key to
be applied. For example we want SingletonStrT keys to be transformed to StrT first.

Changelog: [fix] fixed a source of spurious errors when using refined string keys as computed properties in objects (e.g. [try-Flow](https://flow.org/try/#1N4Igxg9gdgZglgcxALlAJwKYEMwBcD6aArlLnALYYrgA2WAzvXGCADQgYAeOBARgJ74AJhhhYiNXClzEM7DFCLl602QF92kEdRgk8caAAJcGergAUAawz9khgPxm0cKAgCUh4AB0ohw3BhDKxtDAF5Qw0UaGg9vXz9DTFwiNCgAbh8-NUzDSCgzQwheACswz0MAbWt+AF07LxAGwzU0wwB6NsMAdzhcAAtjPoxDeE5WQbh6brhoyIhDGmgEDDRDFbQINB9sqDYQADcVpmhqfYAGADoAJjOANguzkDUgA))

Reviewed By: SamChou19815

Differential Revision: D45938102

fbshipit-source-id: f39378657c88c16fae9cb4a19ec38766f724d53f