Fuse making errors printable with filter suppressed errors
Summary:
With this change, the suppression API now takes a set of Flow_error.t instead of
a set of concretized errors. This API is a nicer starting point for further
optimizations, so future work will build on this.

While we do avoid creating a large set of concrete error messages, this diff
itself does not speed up collate errors very much on its own.

Reviewed By: SamChou19815

Differential Revision: D45916053

fbshipit-source-id: fddf74ab2420027e86ea448edaf4056763b186cf