support refinement with user defined type guards
Summary:
This diff uses the type-guard encoded in the predicate part of function type when calling
`CallLatentPredT`. This allows refinements of the form:
```
declare var fn: (x: mixed) => x is number;
declare var y: mixed;
if (fn(y)) {
  (y: number); // okay
}
```
The refinement of the type happens in Type_filter.intersect, where we compute the intersection
of the type input type with the type guard. For now this function is pretty basic. It uses quick_subytpe
to quickly pick the smaller of the two types, if there is such strict relationship, otherwise it creates
the intersection type. This is a pretty basic formulation of this function. I will be iterating over it in
later diffs.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45659583

fbshipit-source-id: 9daf9368e4fae217df22cd0bf25f18ea8ccb994c