add TypeGuard in Type.fun_predicate
Summary:
This diff adds a `TypeGuardBased` variant to `Type.fun_predicate`. This constructor includes
the parameters name that is refined and the type it is refined to.

These definitions allow the following declarations when `experimental.type_guards` is enabled:
```
type T = (x: A) => x is B;
function f(x: A): x is B { ... };
```
This diff enforce the following checks for these definitions to be well-formed:
* The type guard parameters `x` needs to be a regular parameter of the function (or function type).
* It cannot be a pattern binding or a rest parameter. Both of these cases will result to errors.
The function in general may have patterns or rest params, just not in the predicate spot.
* Type `B` needs to be a subtype of `A`.

Note that the first two checks are implemented in two places for:
* function declarations (statement.ml mk_func_sig)
* function types (type_annotation.ml)

In terms of subtyping, considering
```
type Fa = (x1: S1, ..., xn: Sn) => xi is Ra;
type Fb = (x1: R1, ..., xn: Rn) => xj is Rb;
```
`Fa <: Fb` iff:
* function param types are contravariant as usual
* the type guard position is covariant, so Ra <: Rb
* the type predicates refine the same position in the parameter list, ie. i = j

Also, type-guard based predicate functions are incompatible with predicate-based (`%checks`)
functions, and a non-predicate function is not a subtype of type-guard function.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45629831

fbshipit-source-id: 0ff0b54abba0a4d461542e595c2843bfd998128c