[find-refs] Use the location of imported name instead of the entire import as the reference location
Summary:
Consider the following program for find refs

```
import {a} from './a';
const A = require('./a');
A.a; // <- call find-all-refs on a
```

Currently, the local find all refs will correctly find all refs, but for the import statement, instead of pointing to location of `a`, it will point to the entire import. This looks terrible and this diff fixes it.

Changelog: [ide] In find local references, if a reference appears in a named import, it will no longer highlight the entire import as reference. Instead, it will only highlight the relevant name.

Reviewed By: samwgoldman

Differential Revision: D46033816

fbshipit-source-id: b1a6b9b4bd80fe99c1f2dc800ef55de150d3d064