Add DeclareComponent Support
Summary:
This diff adds support for `declare component()` to Flow. This is part of Flow's component syntax.

It's fairly similar to the standard component syntax but has a few restrictions. Both the name and type annotation are required for parameters but you cannot include a local binding (`as` syntax). Additionally, an identifier is optional for rest parameters. I tried to reuse the code for both parameters and the rest parameter which consolidated code but may have made things a bit more confusing. Let me know if you see anywhere that the code duplication would be preferable.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D45255857

fbshipit-source-id: f8b447794b9f03aab58f3e0fac13796882354bfd