Automatically rename imports on module rename
Summary:
This diff adds the logic to actually make the rename of imports when a module is renamed. It is very bare-bones right now and I've left a number of TODOs in the code. Currently, this feature only works for Haste global imports. It needs to be updated to work with Haste relative requires and Node requires.

Note that this feature is not accessible by the user until the extension update is landed (D44775021). I won't put it up for review until the feature is ready to be launched.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D44229535

fbshipit-source-id: 6521b93dbbd15e829fa54b93becb10ec5e97e13a