[refactor][EZ] Simplify import/export kit returns
Summary:
We can simplify F.return, since in two places that are used, annotation inference always return the type, and flow_js always flow the type to a tout. We can inline these code on the use sites. This helps to eventually move most of the import export logic out of flow_js

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D46171927

fbshipit-source-id: ccf15de40b989eba016243bf1dbf625ebca66a1e