[refactor] Use concretization instead of another flow for importing names
Summary:
Currently, for importing names and importing defaults, we use another flow to validate the imported name. This is needed because the type is usually wrapped in OpenT, TypeT, etc. However, we can do the same with concretization, and simplify the logic.

On the annotation inference side, we have to make the newly added concretization constrain take a transform function, because in annotation inference, we will rewrap the OpenT content with OpenT after the content is elaborated.

With this refactor, the code path for importing names/default can finally return a type directly.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D46175156

fbshipit-source-id: 649b78e4811e6b19487a0d0b3915f91758e1cfa1