[find-all-refs] Gate global identifier and property ref with a single flag
Summary:
I originally think that find all property vs variable refs can go through relatively separate code path and can be rolled out independently, so I give them separate flags.

It turns out that this is false. Consider the test example added in D46033816.

```
import {a} from './a';
const A = require('./a');
a;
A.a;
const AA = {BB:A};
AA.BB.a;
```

When we trigger a property or identifier find all refs, the final results might contain both property and identifiers! Therefore, these two problems are not separate and must be fused into one problem, which I give an initial solution in the next diff.

Meanwhile, let's fuse the flag first.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D46035866

fbshipit-source-id: 5ed509911b6e3354a8e065d0f85e9c9a361e58b2