Do not emit redundant flowlint comments error in libdef
Summary:
Library definitions don't have control over where the configuration where it is used, and it might need to use features that are considered as bad by flow lints (e.g. getters and setters). Flow currently complains about flowlint comments that do nothing, which makes the beforementioned use cases impossible to achieve.

Instead of rejecting this, which is too strict, or do not emit any lint errors in libdef, which is too loose, we need to at least let the users know what they are doing and give them a way to shut Flow up, which is what's being done in this diff.

Changelog: [fix] We will no longer emit redundant flowlint disable errors in library definitions.

Reviewed By: williewillus

Differential Revision: D46165188

fbshipit-source-id: a4c1544876fe3ef9dafb2e68dee6dbb95e5165fc