De-functorize `ImportTypeTKit` and `ImportTypeofTKit`
Summary:
With previous refactors, we cut of a lot of dependency on `Import_export_helper_sig` for these two kits. Instead of relying on a few functions in it, we can just make the caller pass in the required functions.

This function opens up the usage of these kits in statement.ml, and the potential to kill `ImportTypeT` and `ImportTypeofT`.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D46205969

fbshipit-source-id: 70b967d8d3a03105ec117a1364bd1d996b7dda46