[get-def] Make get-def reliably jump to the definition of exported name
Summary:
This diff starts to take advantage of the reliable definition location of name stored in typed ast, to ensure that get def for esm exports always jump to the name location. This is necessary to make get-def reliably power find all refs, it is also a more principled response to get-def: it starts from a name location, so it should end at a name location.

Changelog: [ide] Get definition request will now consistently jump to the name of the definition for ES module exports.

- For exports like `export const a = ...`/ or `export default function a() {...}`, it will always jump to `a`
- For exports like `const a = ...; export {a}`, it will always jump to `a` in `const a = ...`
- For exports like `const a = ...; export default a`, it will always jump to `a` in `const a = ...`
- For exports like `export default <some arbitrary expression>`, it will always jump to location of `default`

Reviewed By: panagosg7

Differential Revision: D46211929

fbshipit-source-id: cd979e072c73992817e5e443c721156794592cd6