[scope] fix parameter redeclaration in func body when some parameter has defaults
Summary:
Today, the following code
```
function fn(x: number, y: number = 1) {
  let x = 0;
}
```
is not an error, even though this fails at runtime with
```
Uncaught SyntaxError: Identifier 'x' has already been declared
```
The reason is that when we're dealing with a function body in the scope builder under
`has_default_parameters = true`, we are visiting the body of a function in a scope different
than the one that we visited the params. Therefore, declarations in the body of the function
are see for the first time, and so a let-declaration on the same name with a variable is considered
okay.

This diff addresses this by first visiting the default expressions in their own scope, as they should
be since bindings from the function body should not be visible, and then by visiting the parameters
again in the scope of the function body, but this time without the defaults.

Changelog: [fix] report error when a function parameter is redeclared in its body in functions that have default params (e.g. [try-Flow](https://flow.org/try/#1N4Igxg9gdgZglgcxALlAJwKYEMwBcD6aArlLnALYYrgA2WAzvXGCADQgYAeOBARgJ74AJhhhYiNXClzEM7DFCLl602QF92kEdRgk8caAAIYUABSdkhxeV4Y0rQ-0vXbaQwF5DARgCUh4AA6UIaGNBi4hpwehgAMANyGAPSJhgDucDQ0VhCphraGWMF2aBBBamwgAG52TNDUlTEAdABMMQBsjTEgakA))

Reviewed By: SamChou19815

Differential Revision: D46093690

fbshipit-source-id: 895ab65c87520aa8a12e4bfa90c49a6cafdcbe51