[scope] param types may reference names appearing earlier in param list
Summary:
Function parameters used to not be able to refer to earlier parameters. Ie the following was an error
```
function f(x: number, y: typeof x) {} // error
```
This was implemented in D35478747 with the understanding that when new envs and LTI would roll out,
this behavior would be restored to the more straight-forward case where parameter types can refer to
other parameters defined earlier. This is also consistent with default vales, that *can* refer to earlier params:
```
function f(
    x: number,
    y: number = x, // x refers to the param above
) {}
```

This diff does this by effectively reverting D35478747.

It then folds over the parameters of a lambda in Scope_builder, and visits each annotation
in an environment that includes all the bindings of the previous params.

Changelog: [fix] parameter types can now refer to previous parameters in the same parameter list (e.g. `function f(x: number, y: typeof x) {}`)

Reviewed By: jbrown215

Differential Revision: D46165727

fbshipit-source-id: 6bd14a0b7c59cac14918a5328add95cb174de167