Parse negative number/bigint literal types as two tokens
Summary:
Instead of trying to deal with negative literal types in the lexer, emit a
T_MINUS token separately and deal with it in the parser.

This change has a few observable differences:

1. The `raw` field does not include any whitespace between the `-` and the
number/bigint.

2. Errors for legacy octal literal in strict mode point to the number part only,
excluding the `-`.

3. Errors for octal bignumb literals (before this diff and after, this is an
illegal token error) point to the number part only, excluding the `-`.

I think these are minor. For (2) and (3), the changed behavior mirrors the
behavior for values, where the `-` is actually a separate AST node for the unary
minus operator.

Reviewed By: mroch

Differential Revision: D46007203

fbshipit-source-id: 66a3d45a4a84eea3e97b8d9543a284129602a8f1