[find-all-refs] Initial global find variable refs support
Summary:
After all the prep diff before, we are ready to implement global find variables refs.

We first consider how we can find all the variable references to a variable. Given that we know a definition location, a reference is either inside the same file or outside the file. For the one inside the file, a local find ref based on scope builder should be enough. If it's in another file, then the variable must first be imported. We will inspect all the imports, find all the local bindings that corresponds to the original def loc (computed by the newly created `LocalImportRefSearcher` module), and then run a scope builder based find all refs.

As described in the previous diff, the result has to be fused together with property def. If we initially start with property find ref, then we can just use all the def locs and run a variable find ref as well. Instead, if we start from variable find refs, we can consider all the def locs as `ObjectProperty` for find property refs, since find property refs model module exports as `ObjectProperty`.

With this, we have an initial prototype for all kinds of find refs working!

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46036597

fbshipit-source-id: 67c99d3c3b3274cca8360782f9207d394d809d98