[component-syntax] Add tests demonstrating erroneous ssa_builder behavior with components
Summary:
The ssa builder does not handle abrupt completions in components correctly yet, leading it to erroneously report dead reads. In the added tests you can see how we report the read of a function after its body but do not report the read of the component after its body-- this is because the `return` is not caught and bubbles up to the program scope, so the ssa builder never even reads that line.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D46322890

fbshipit-source-id: 95ddb446910f5cb69eeb5a644ef445edfb77e719