[component-syntax] Fix ssa-builder handling for components
Summary: The ssa_builder informs the name_resolver about which names are "unbound" and should be treated as globals. However, the SSA builder did not have a case to handle the abrupt completion scope of a component, which means that a `return` inside a component would bubble up higher than the component declaration, leading to strange results. This diff adds the same handling for functions to components, which fixes the test added in the previous diff

Reviewed By: SamChou19815

Differential Revision: D46322892

fbshipit-source-id: f60ced85634475fd01424d40912fd31e2bb7b060