[component-syntax] Introduce component_sig and friends
Summary:
This diff introduces the type checking code that will handle components. This is not the most obvious approach, so allow me to explain:

We plan to support declared components, component types, component declarations, and possibly component expressions. All of these variants of a component will have a different AST structure, which means type checking code would have to be written for each if we approached this problem naively. Luckily, we've already handled problems like these with functions, so we can take inspiration from how it's handled there.

The overall approach here is to define a common interfaces for parameters and component signatures and then write adapters to convert each concrete AST representation into our our common representation. We then provide a module to check component bodies and determine the config type given this common representation.

This diff adds the following modules:
1. component_sig_types: Defines types to represent component signatures and parameters.  Importantly, we use this module in statement_sig.ml so we cannot include it in component_sig or we would create a cycle.
2. component_params: this module is responsible for generating the config type of a component given params in the common representation. It also constructs the typed asts for those things. Same constraints about statement_sig.ml and cycles as described above.
3. component_sig_intf: This defines the signature for component_sig. It's only needed to ensure we do not create any cycles with statement_sig.ml
4. component_declaration_config: the concrete implementation for turning component declarations into the common representation
5. component_sig: given a a component common representation, checks the body and determines the signature

This mostly mirrors the func_sig support, but cuts down on some interface files.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D45876626

fbshipit-source-id: ccce77993e8c0c8b6c3fadd1f122969e1b1f546a