disallow variable annotation to reference directly the variable name
Summary:
We now disallow the annotation on a variable declarator to reference directly the variable name being declared.

This means that declarations like
```
const x: typeof x = ...
const a: Array<typeof a> = ...
```
are now invalid.

Most of these patterns are already problematic in that they cause errors due to the side-effect of
being resolved to the `empty` type. For example
```
const x: typeof x = 1;
```
errored with
```
1: const x: typeof x = 1;
                       ^ Cannot assign `1` to `x` because number [1] is incompatible with empty [2]. [incompatible-type]
References:
1: const x: typeof x = 1;
                       ^ [1]
1: const x: typeof x = 1;
            ^ [2]
```

This change helps us prevent two more problematic scenarios:

1. A bad error message in `const x: x = 1;`, where again we get "number is incompatible with empty",
instead of value-as-type.
2. A recursion-limit-exceeded crash on
```
declare var x: x<>;
const {f} = x;
```

**Exceptions**

It will still be possible to refer to the binding being defined when the reference is under an object type
constructor, given that this is a common way of defining recursive types, and does not have the short-
comings of the top-level references. E.g.
```
const o: { f: typeof o } = ...
```

It will also still be possible to define a recursive type through type annotation:
```
type R = R | null;
```
though these types remain problematic.

**Implementation**

The check is implemented in the name_resolver, we keep track of the currently processed parameter
def-locs, so that we can error if there exist references to these names in the corresponding type annotation.

Changelog: [fix] we now error when using the name of a variable in its own annotation (e.g. `const x: typeof x = ...`). An exception is when the name of the variable appears under an object type constructor. For example, the annotation `const o: { f: typeof o } = ...` is allowed.

Reviewed By: SamChou19815

Differential Revision: D46006084

fbshipit-source-id: a4236a211e7109909e02b7e1051c502899e42af5