[component-syntax] Typecheck component params and bodies
Summary:
This diff implements `mk_component_sig`, which takes in a component declaration and outputs the common data type used in the machinery provided by the previous diff. We can now check the body and params of a component.

An important point in this transformation is that we assume a React$Node annotation return type when no return annotation is provided.

Reviewed By: SamChou19815

Differential Revision: D45876624

fbshipit-source-id: 211ccd85e146a3f6ca09c8677874543b9c54fd7c