Get rid of "literal" string for tagged template token
Summary:
For tagged templates, we stored three different kinds of strings: (a) cooked/value,
which is the contents of the template with any escape codes decoded; (b) raw,
which is the contents of the template with the original escape codes, but
without the delimeters, and (c) literal, which is the same as raw, but including
delimiters.

This diff elimiates the "literal" data, which can be trivially reconstructed
from "raw" by keeping track of whether the template part is the "head" and
"tail" of a template literal string.

This change is a step toward unifying string literal and template literal
lexing. String lexing has only raw and decoded variants, so removing this third
value makes it easier to combine code paths.

(Naturally, this is more efficient as well, but lexing template literal string
parts is hardly a bottleneck.)

Reviewed By: mroch

Differential Revision: D46286330

fbshipit-source-id: 3e2ed23a039bc1c7fbcb3e31c22b41afa92ecc5f