[component-syntax] Visit component defaults and annotations explicitly without the params in scope
Summary:
The param bindings in a component are not in scope for the annotations/defaults of the component. Prior to this diff, we visited the entire params, including the binding identifiers, without putting the params in scope. This leads to the ssa_builder reporting them as unbound names in the environment. With that, we initialize them in the environment as a global, so they will always be in the env_state, and when we see the parameter bindings we set their write state to the parameter locations. This makes these params behave as if they are initialized when visiting the defaults.

This diff visits the defaults and type annotations explicitly without the bindings and then visits only the params and not the annotations/defaults with the param bindings in scope. This prevents the ssa_builder from reporting them as globals and correctly keeps them out of scope when visiting the defaults and annotations.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D46319795

fbshipit-source-id: 0b26dd033544bb310a73b15fdc23d7fe6b489541