Ban redeclaration of props and "ref" props everywhere
Summary:
This diff builds the config type for a component using a new object kit tool, replacing the use of spread originally. This tool validates that there is no overlap between inline properties and spread properties, and errors if such overlap exists. (It is also substantially simpler than the spread tool, because by design we never need to worry about multiple spreads, merging properties, or dictionaries on both sides.)

One subtle bit of the specification is the behavior of indexers. In some sense, allowing indexers like `{[string]: T}` in spread props when there are any inline props is a violation of the specification: `component (bar: number, ...props: { [string]: boolean }) { }` kinda has "bar" in both the inline and spread props. However, because we know that the inline props are fully specified and there's no "inline indexer prop", we can allow this and produce a config that looks more like `{[string]: boolean, bar:number}`, i.e. the inline property overrides the indexer. This is the tack that this diff takes, since to do otherwise would effectively prevent indexers from existing in most cases.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D46359525

fbshipit-source-id: e114e4a1fab1db8b1cd5744dfccbfed08f1e9259