Implementation of component type sigs
Summary:
This diff implements type sigs for component syntax: in particular, it implements parsing and merging components.

The implementation here mostly closely follows the implementation for functions, especially on the parsing side. We do pass through some additional locations, such as the overall parameter location, which necessitates taking care to preserve the type_sig ordering invariants.

On the merging side, the main difference here is that we need to combine all the parameters into a config object. We do so using the same destructor that we use for typechecking components, which consumes a property map and optionally a spread argument and produces a single argument for the resulting FunT.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46368444

fbshipit-source-id: 380d293793719e5392675980afad6f52a226b68e