[component-syntax] Separate Config module type into Component_params_intf module
Summary:
The `Config` module type is the only one in `Component_sig_types` that references `Context`. In the next diff, I start to use `Component_declaration_sig_types.t` in the component_sig node cache. Importantly, the node cache module cannot depend on Context because the node cache is stored in Context, which is why I need to separate out the Config module in this diff.

This diff follows the `func_params_intf.ml` module, which pulled out the config definition into its own file.

Most of the changes here are renaming references to the Config module and otherwise are just moving Config to a new file.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D46370959

fbshipit-source-id: bfb0183a6ccf362bf9e7a7b48b7b26005e9f7d67