[get-def] Return error results for named imports without populated remote def_loc
Summary:
As long as the import is legit, these remote def locs should be reliably populated, since the information is from type_sig with the guarantee that it points to the name.

If there are issues with the def loc from type sig, we should fix type sig and its populating code rather than relying on this fallback that relies on poorly defined def_loc spec in reasons.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46344372

fbshipit-source-id: ecc6ac0b064604e6e3d46b2fd6fdeb357e5e7657