Refine literal autocomplete to just string literals
Summary:
Currently, we do the autocomplete hooks on all literals, but currently
autocomplete is only reasonable for strings.

Recall how literal autocomplete works today:
1. We insert the sigil "AUTO332" after the cursor in the raw contents and parse
2. We get the typed AST for the modified file, and find the cursor in the typed AST
3. When the cursor is on a literal AST node, we look at the upper bounds to get completions
4. The possible completions are filtered by the prefix, if any

Now consider completing `var x: T = 4|`. We would modify that to `var x: T = 4AUTO332`,
which is a parse error. The parser currently recovers by ignoring the `4` and
creating an identifier node, `var x: T = AUTO332`.

The issue above is fixable, but the upshot here is that literal autocomplete
currently only works for strings. This diff changes the code to better reflect
this current reality.

This is further reinforced by our tests. Making this change has no effect on our
tests because the behavior for non-string literals is untested. Indeed, if the
tests were written, they would not exhibit the expected behavior!

Why do this? My underlying motivation is unrelated to autocomplete. I want to
change the representation of literals in the AST. Instead of all literal nodes
sharing the same encoding, I want to break them out. This makes it possible to
optimize parser performance and memory use of the AST for strings.

Reviewed By: SamChou19815

Differential Revision: D46400970

fbshipit-source-id: 5c8239b88f6944c34175bbde69a723d35e7380fc