[component-syntax] Disallow Rebinding/Reassigning components
Summary:
Components should not be re-bound nor reassigned. This diff introduces a new component binding in bindings.ml that we use in the name_resolver to detect and emit errors when components are rebound/reassigned.

It's also important to test that other bindings that do not allow reassignments cannot be reassigned by components, which is also included in the added test.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D46376098

fbshipit-source-id: 980e15b7a0dfa67f9171f294c33029823d93aeba