[component-syntax] Do not require annotations on functions due to nested components
Summary:
We only require annotations on exported function returns if they are not a procedure. The `Procedure_decider` modules determines if a function is a procedure by looking for return statements. Notably, it does not recurse into nested function declarations, as those returns do not reach the exported function's signature. We should similarly skip descending into components.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46405777

fbshipit-source-id: 5bdc4d65971cb2fd7de0ac8b0b96c5eba84bef16