Provider support for components
Summary:
This diff builds support for component declarations in the find_providers module. This work is necessary in order to fit components into the algorithm for finding providers for assigned variables--e.g., recognizing that assignments within component bodies are within nested scopes, and therefore assignments below it in shallower scopes take priority. For example, in the added test case, without this diff we would consider the assignment inside the component to take priority, while with this diff we take the lower assignment.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D46455425

fbshipit-source-id: dd104daeaf03f30389c7cd74d43690bee6b700b4