Break up literal ast nodes
Summary:
Changing the AST is kind of a nightmare, and reviewing this diff might also be.
Sorry :(

Instead of a single Expression.Literal constructor with a variant, this diff
uses separte constructors for different kinds of literals. This is how most
ASTs model this: Babel, Hermes, TypeScript.

The Flow AST is like this because estree is like this, but this way is better in
almost every way.

Importantly, it makes it nicer to have different (or no) "raw" representations
for different types. Bool and Null literals don't need raw representations. For
strings, we'd like to avoid storing the raw-with-quotes representations to save
space.

The changes here are extensive, but I worked to maintain existing behavior,
including bugs I noticed on the way. (Mostly these bugs are mis-using "raw" for
strings, which includes quotes.)

In follow-up diffs, I'll address the bugs I found, but mainly exploit this
change to optimize parsing code and memory use for the AST, with knock-on perf
benefits for merge and check phases.

Reviewed By: SamChou19815

Differential Revision: D46414261

fbshipit-source-id: cf1ec644d83fc5b98d19dc0e7bbcbdfa496e0825