Delete proc_test
Summary:
D39490515 deletes the aa_injection_hack, but this test I propose to delete in this diff actually depends on the hack to use mocked disk, otherwise it might fail with permission error, like we see today.

The test has been disabled internally for quite a long time, and it's never run in the OSS. The code is quite stable, so we should just delete the test, unless someone wants to bring back the mock just for this single test.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46448816

fbshipit-source-id: 8fda17898590580c4a43c6b438481dc6382c31f2