Ban implicit returns in components
Summary:
This diff bans implicit returns in the component syntax. The bulk of this work is in exhaustiveness checking: just as we don't flow void to the return of a function if the the function returns in all cases of an exhaustive switch, we also don't error if a component returns in all such cases. This requires us to implement a (simplified) equivalent of the function exhaustiveness handling work, with the change that rather than propagating void to the return tvar if a lower bound reaches ImplicitVoidReturn, we now directly error in components when this happens.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D46462113

fbshipit-source-id: a55dfe2955bb2a5969120ba0cfded50513f26fb4