[get-def] Get rid of TypeT special casing
Summary:
This was originally added in D23435319, but none of these cases will be hit now. For named imports, we will use def_loc from type sig. For namespace imports, we will directly jump to source without the indirection of import types. Therefore, we can delete this to make the analysis easier to understand.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46416831

fbshipit-source-id: 85c81c38a87ba1686def039e7eb7253566d3f07d