remove duplicate error on non-returning predicate function
Summary:
Currently we would return two errors for the code
```
function no_return(x: mixed): x is number {}
```
An [incompatible-return] due to undefined ~> boolean, and a special [unsupported-syntax] error for predicate functions.

This diff removes the second error, and customizes the first message when applied to a predicate function.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D46449291

fbshipit-source-id: 669685113c708bfdd3372f0ba54a307519d0482f