move predicate func checking to the end of Statement.mk_func_sig
Summary:
This diff moves predicate related checks, that are now part of mk_func_sig, to happen after the body of the function has been checked. This change ensures that the body has been properly type checked and the envs are up-to-date, before we attempt to raise any predicate/type guard related errors.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D46449292

fbshipit-source-id: 4e94ca80204078de6eeb6628958025bcd01404a0