skip type guard param in name-def ordering
Summary:
This diff adds two tests on name-def ordering where a parameter name appears in the return annotation
```
declare var f: any;
f((x): typeof x => x);
f((x): x is number => true);
```
Without any intervention, and due to the previous diff, both of these tests would raise an error due to an illegal self-cycle.

Neither of these cases are answered in the codebases. The first, because until the previous diff, `x` was not in scope in the return annotation, and the second because type guards have not been released yet.

This diff addresses the cycle issue in the case of `x is T`, by skipping the parameter name during name-def-ordering. This won't actually add any useful information anyway, since are not interested in the type of parameter `x`, just its position in the parameter list.

The first case is more subtle, but since it's extremely rare, I'm leaving as future work. The workaround of course is to add an annotation to `x` or a different type as return, both of which are very reasonable.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D46416116

fbshipit-source-id: 0610e67f5d9cd414960845ee45ea0f922d7db562