[get-def] Get rid of AnnotT special casing
Summary:
This was originally added in D22464315, but now it's no longer required. I believe it's required because at the time of the diff, it will create AnnotT that mess up results:

https://www.internalfb.com/code/fbsource/[94f8ab1fa34ecdec82230ac32c2d6ab8c59cf898]/fbcode/flow/src/typing/flow_js.ml?lines=11894-11905

, where mk_instance is used to create refs in type_sig a lot.

However, nowadays type_sig will use annotation inference module to generate these types, with much simpler reason positioning logic, so the original assumption is no longer true. We should kill this codepath to make the code easier to understand.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46416830

fbshipit-source-id: 3946c63bef065849a9bb3b36023c69a6f83b92da