Support multi-file lint suppression cover
Summary:
I refactored the lint suppression cover code to support the use case of generating a cover for multiple files.

It works simply by combining single file results into a map. It partially reverts the change in D46283718 that tries to do it together, which doesn't work due to the assumption that the processing code works for a single file.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D46448204

fbshipit-source-id: edc322cdbc5b1f9c0f3cfa5cb82be577b8db624e