[component-syntax] Rename return to renders for components
Summary:
Right now we use an annotation to describe the return type of a component:
```
component Foo(): RETURN { ... }
```

We're moving away from this design to encourage people to only use the "return" position of a component when describing what a component renders. The new syntax is going to be:
```
component Foo() renders RENDERSTYPE
```

This diff changes the AST fields to `renders` instead of `return`, along with the Estree output. This diff does *not* change the parsing rules, which will be done in a stacked diff.

Reviewed By: SamChou19815

Differential Revision: D46509471

fbshipit-source-id: 38b1e7046fdfd227605b5999be08a44f9f28774f